refactor: split cli into modules
